pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "caption-extractor"
dynamic = ["version"]
description = "OCR text extraction from images using PaddleOCR PP-OCRv5"
readme = "README.md"
license = "MIT"
requires-python = ">=3.8"
authors = [
    {name = "Caption Extractor Team"},
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
]

dependencies = [
    "paddlepaddle>=2.5.0",
    "paddleocr>=2.7.0",
    "pyyaml>=6.0",
    "pillow>=10.0.0",
    "opencv-python-headless>=4.8.0",
    "tqdm>=4.64.0",
    "numpy>=1.21.0,<2.0.0",
    "protobuf>=3.20.0,<5.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
    "mypy>=1.0.0",
]

[project.urls]
Homepage = "https://github.com/devopsnextgenx/caption-extractor"
Repository = "https://github.com/devopsnextgenx/caption-extractor.git"
Issues = "https://github.com/devopsnextgenx/caption-extractor/issues"

[project.scripts]
caption-extractor = "caption_extractor.main:main"

[tool.hatch.version]
path = "src/caption_extractor/__init__.py"

[tool.hatch.build.targets.wheel]
packages = ["src/caption_extractor"]

[tool.black]
line-length = 88
target-version = ['py38']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --tb=short"